refactored sampler operations into a single external file
