+++ packages/contract-wrappers/src/generated-wrappers/erc20_bridge_sampler.ts
@@ -33,8 +33,6 @@ import { assert } from '@0x/assert';
 import * as ethers from 'ethers';
 // tslint:enable:no-unused-variable
 
-
-
 /* istanbul ignore next */
 // tslint:disable:array-type
 // tslint:disable:no-parameter-reassignment
@@ -43,14 +41,14 @@ export class ERC20BridgeSamplerContract extends BaseContract {
     /**
      * @ignore
      */
-public static deployedBytecode: string | undefined;
-public static contractName = 'ERC20BridgeSampler';
+    public static deployedBytecode: string | undefined;
+    public static contractName = 'ERC20BridgeSampler';
     private readonly _methodABIIndex: { [name: string]: number } = {};
-public static async deployFrom0xArtifactAsync(
+    public static async deployFrom0xArtifactAsync(
         artifact: ContractArtifact | SimpleContractArtifact,
         supportedProvider: SupportedProvider,
         txDefaults: Partial<TxData>,
-        logDecodeDependencies: { [contractName: string]: (ContractArtifact | SimpleContractArtifact) },
+        logDecodeDependencies: { [contractName: string]: ContractArtifact | SimpleContractArtifact },
     ): Promise<ERC20BridgeSamplerContract> {
         assert.doesConformToSchema('txDefaults', txDefaults, schemas.txDataSchema, [
             schemas.addressSchema,
@@ -69,7 +67,13 @@ public static async deployFrom0xArtifactAsync(
                 logDecodeDependenciesAbiOnly[key] = logDecodeDependencies[key].compilerOutput.abi;
             }
         }
-        return ERC20BridgeSamplerContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly, );
+        return ERC20BridgeSamplerContract.deployAsync(
+            bytecode,
+            abi,
+            provider,
+            txDefaults,
+            logDecodeDependenciesAbiOnly,
+        );
     }
 
     public static async deployWithLibrariesFrom0xArtifactAsync(
@@ -77,7 +81,7 @@ public static async deployFrom0xArtifactAsync(
         libraryArtifacts: { [libraryName: string]: ContractArtifact },
         supportedProvider: SupportedProvider,
         txDefaults: Partial<TxData>,
-        logDecodeDependencies: { [contractName: string]: (ContractArtifact | SimpleContractArtifact) },
+        logDecodeDependencies: { [contractName: string]: ContractArtifact | SimpleContractArtifact },
     ): Promise<ERC20BridgeSamplerContract> {
         assert.doesConformToSchema('txDefaults', txDefaults, schemas.txDataSchema, [
             schemas.addressSchema,
@@ -99,13 +103,16 @@ public static async deployFrom0xArtifactAsync(
             artifact,
             libraryArtifacts,
             new Web3Wrapper(provider),
-            txDefaults
+            txDefaults,
         );
-        const bytecode = linkLibrariesInBytecode(
-            artifact,
-            libraryAddresses,
+        const bytecode = linkLibrariesInBytecode(artifact, libraryAddresses);
+        return ERC20BridgeSamplerContract.deployAsync(
+            bytecode,
+            abi,
+            provider,
+            txDefaults,
+            logDecodeDependenciesAbiOnly,
         );
-        return ERC20BridgeSamplerContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly, );
     }
 
     public static async deployAsync(
@@ -123,11 +130,7 @@ public static async deployFrom0xArtifactAsync(
         ]);
         const provider = providerUtils.standardizeOrThrow(supportedProvider);
         const constructorAbi = BaseContract._lookupConstructorAbi(abi);
-        [] = BaseContract._formatABIDataItemList(
-            constructorAbi.inputs,
-            [],
-            BaseContract._bigNumberToString,
-        );
+        [] = BaseContract._formatABIDataItemList(constructorAbi.inputs, [], BaseContract._bigNumberToString);
         const iface = new ethers.utils.Interface(abi);
         const deployInfo = iface.deployFunction;
         const txData = deployInfo.encode(bytecode, []);
@@ -143,7 +146,12 @@ public static async deployFrom0xArtifactAsync(
         logUtils.log(`transactionHash: ${txHash}`);
         const txReceipt = await web3Wrapper.awaitTransactionSuccessAsync(txHash);
         logUtils.log(`ERC20BridgeSampler successfully deployed at ${txReceipt.contractAddress}`);
-        const contractInstance = new ERC20BridgeSamplerContract(txReceipt.contractAddress as string, provider, txDefaults, logDecodeDependencies);
+        const contractInstance = new ERC20BridgeSamplerContract(
+            txReceipt.contractAddress as string,
+            provider,
+            txDefaults,
+            logDecodeDependencies,
+        );
         contractInstance.constructorArgs = [];
         return contractInstance;
     }
@@ -153,7 +161,7 @@ public static async deployFrom0xArtifactAsync(
      */
     public static ABI(): ContractAbi {
         const abi = [
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -172,7 +180,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -199,7 +207,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -262,7 +270,7 @@ public static async deployFrom0xArtifactAsync(
                                 name: 'takerFeeAssetData',
                                 type: 'bytes',
                             },
-                        ]
+                        ],
                     },
                     {
                         name: 'orderSignatures',
@@ -280,7 +288,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -343,7 +351,7 @@ public static async deployFrom0xArtifactAsync(
                                 name: 'takerFeeAssetData',
                                 type: 'bytes',
                             },
-                        ]
+                        ],
                     },
                     {
                         name: 'orderSignatures',
@@ -361,7 +369,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -388,7 +396,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -419,7 +427,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -446,7 +454,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -477,7 +485,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -504,7 +512,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -531,7 +539,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -562,7 +570,7 @@ public static async deployFrom0xArtifactAsync(
                 stateMutability: 'view',
                 type: 'function',
             },
-            { 
+            {
                 constant: true,
                 inputs: [
                     {
@@ -619,10 +627,7 @@ public static async deployFrom0xArtifactAsync(
                         libraryAddresses,
                     );
                     // Deploy this library.
-                    const linkedLibraryBytecode = linkLibrariesInBytecode(
-                        libraryArtifact,
-                        libraryAddresses,
-                    );
+                    const linkedLibraryBytecode = linkLibrariesInBytecode(libraryArtifact, libraryAddresses);
                     const txDataWithDefaults = await BaseContract._applyDefaultsToContractTxDataAsync(
                         {
                             data: linkedLibraryBytecode,
@@ -673,485 +678,481 @@ public static async deployFrom0xArtifactAsync(
 
     /**
      * Call multiple public functions on this contract in a single transaction.
-      * @param callDatas ABI-encoded call data for each function call.
-    * @returns callResults ABI-encoded results data for each call.
+     * @param callDatas ABI-encoded call data for each function call.
+     * @returns callResults ABI-encoded results data for each call.
      */
-    public batchCall(
-            callDatas: string[],
-    ): ContractFunctionObj<string[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isArray('callDatas', callDatas);
+    public batchCall(callDatas: string[]): ContractFunctionObj<string[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isArray('callDatas', callDatas);
         const functionSignature = 'batchCall(bytes[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<string[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<string[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<string[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<string[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [callDatas
-            ]);
+                return self._strictEncodeArguments(functionSignature, [callDatas]);
             },
-        }
-    };
+        };
+    }
     /**
      * Returns the address of a liquidity provider for the given market
- * (takerToken, makerToken), from a registry of liquidity providers.
- * Returns address(0) if no such provider exists in the registry.
-      * @param takerToken Taker asset managed by liquidity provider.
-      * @param makerToken Maker asset managed by liquidity provider.
-    * @returns providerAddress Address of the liquidity provider.
+     * (takerToken, makerToken), from a registry of liquidity providers.
+     * Returns address(0) if no such provider exists in the registry.
+     * @param takerToken Taker asset managed by liquidity provider.
+     * @param makerToken Maker asset managed by liquidity provider.
+     * @returns providerAddress Address of the liquidity provider.
      */
     public getLiquidityProviderFromRegistry(
-            registryAddress: string,
-            takerToken: string,
-            makerToken: string,
-    ): ContractFunctionObj<string
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isString('registryAddress', registryAddress);
-            assert.isString('takerToken', takerToken);
-            assert.isString('makerToken', makerToken);
+        registryAddress: string,
+        takerToken: string,
+        makerToken: string,
+    ): ContractFunctionObj<string> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isString('registryAddress', registryAddress);
+        assert.isString('takerToken', takerToken);
+        assert.isString('makerToken', makerToken);
         const functionSignature = 'getLiquidityProviderFromRegistry(address,address,address)';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<string
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<string> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<string
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<string>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [registryAddress.toLowerCase(),
-            takerToken.toLowerCase(),
-            makerToken.toLowerCase()
-            ]);
+                return self._strictEncodeArguments(functionSignature, [
+                    registryAddress.toLowerCase(),
+                    takerToken.toLowerCase(),
+                    makerToken.toLowerCase(),
+                ]);
             },
-        }
-    };
+        };
+    }
     /**
      * Queries the fillable taker asset amounts of native orders.
- * Effectively ignores orders that have empty signatures or
-      * @param orders Native orders to query.
-      * @param orderSignatures Signatures for each respective order in `orders`.
-    * @returns orderFillableMakerAssetAmounts How much maker asset can be filled         by each order in &#x60;orders&#x60;.
+     * Effectively ignores orders that have empty signatures or
+     * @param orders Native orders to query.
+     * @param orderSignatures Signatures for each respective order in `orders`.
+     * @returns orderFillableMakerAssetAmounts How much maker asset can be filled         by each order in &#x60;orders&#x60;.
      */
     public getOrderFillableMakerAssetAmounts(
-            orders: Array<{makerAddress: string;takerAddress: string;feeRecipientAddress: string;senderAddress: string;makerAssetAmount: BigNumber;takerAssetAmount: BigNumber;makerFee: BigNumber;takerFee: BigNumber;expirationTimeSeconds: BigNumber;salt: BigNumber;makerAssetData: string;takerAssetData: string;makerFeeAssetData: string;takerFeeAssetData: string}>,
-            orderSignatures: string[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isArray('orders', orders);
-            assert.isArray('orderSignatures', orderSignatures);
-        const functionSignature = 'getOrderFillableMakerAssetAmounts((address,address,address,address,uint256,uint256,uint256,uint256,uint256,uint256,bytes,bytes,bytes,bytes)[],bytes[])';
+        orders: Array<{
+            makerAddress: string;
+            takerAddress: string;
+            feeRecipientAddress: string;
+            senderAddress: string;
+            makerAssetAmount: BigNumber;
+            takerAssetAmount: BigNumber;
+            makerFee: BigNumber;
+            takerFee: BigNumber;
+            expirationTimeSeconds: BigNumber;
+            salt: BigNumber;
+            makerAssetData: string;
+            takerAssetData: string;
+            makerFeeAssetData: string;
+            takerFeeAssetData: string;
+        }>,
+        orderSignatures: string[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isArray('orders', orders);
+        assert.isArray('orderSignatures', orderSignatures);
+        const functionSignature =
+            'getOrderFillableMakerAssetAmounts((address,address,address,address,uint256,uint256,uint256,uint256,uint256,uint256,bytes,bytes,bytes,bytes)[],bytes[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [orders,
-            orderSignatures
-            ]);
+                return self._strictEncodeArguments(functionSignature, [orders, orderSignatures]);
             },
-        }
-    };
+        };
+    }
     /**
      * Queries the fillable taker asset amounts of native orders.
- * Effectively ignores orders that have empty signatures or
- * maker/taker asset amounts (returning 0).
-      * @param orders Native orders to query.
-      * @param orderSignatures Signatures for each respective order in `orders`.
-    * @returns orderFillableTakerAssetAmounts How much taker asset can be filled         by each order in &#x60;orders&#x60;.
+     * Effectively ignores orders that have empty signatures or
+     * maker/taker asset amounts (returning 0).
+     * @param orders Native orders to query.
+     * @param orderSignatures Signatures for each respective order in `orders`.
+     * @returns orderFillableTakerAssetAmounts How much taker asset can be filled         by each order in &#x60;orders&#x60;.
      */
     public getOrderFillableTakerAssetAmounts(
-            orders: Array<{makerAddress: string;takerAddress: string;feeRecipientAddress: string;senderAddress: string;makerAssetAmount: BigNumber;takerAssetAmount: BigNumber;makerFee: BigNumber;takerFee: BigNumber;expirationTimeSeconds: BigNumber;salt: BigNumber;makerAssetData: string;takerAssetData: string;makerFeeAssetData: string;takerFeeAssetData: string}>,
-            orderSignatures: string[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isArray('orders', orders);
-            assert.isArray('orderSignatures', orderSignatures);
-        const functionSignature = 'getOrderFillableTakerAssetAmounts((address,address,address,address,uint256,uint256,uint256,uint256,uint256,uint256,bytes,bytes,bytes,bytes)[],bytes[])';
+        orders: Array<{
+            makerAddress: string;
+            takerAddress: string;
+            feeRecipientAddress: string;
+            senderAddress: string;
+            makerAssetAmount: BigNumber;
+            takerAssetAmount: BigNumber;
+            makerFee: BigNumber;
+            takerFee: BigNumber;
+            expirationTimeSeconds: BigNumber;
+            salt: BigNumber;
+            makerAssetData: string;
+            takerAssetData: string;
+            makerFeeAssetData: string;
+            takerFeeAssetData: string;
+        }>,
+        orderSignatures: string[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isArray('orders', orders);
+        assert.isArray('orderSignatures', orderSignatures);
+        const functionSignature =
+            'getOrderFillableTakerAssetAmounts((address,address,address,address,uint256,uint256,uint256,uint256,uint256,uint256,bytes,bytes,bytes,bytes)[],bytes[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [orders,
-            orderSignatures
-            ]);
+                return self._strictEncodeArguments(functionSignature, [orders, orderSignatures]);
             },
-        }
-    };
+        };
+    }
     /**
      * Sample buy quotes from Eth2Dai/Oasis.
-      * @param takerToken Address of the taker token (what to sell).
-      * @param makerToken Address of the maker token (what to buy).
-    * @returns takerTokenAmounts Taker amounts sold at each maker token         amount.
+     * @param takerToken Address of the taker token (what to sell).
+     * @param makerToken Address of the maker token (what to buy).
+     * @returns takerTokenAmounts Taker amounts sold at each maker token         amount.
      */
     public sampleBuysFromEth2Dai(
-            takerToken: string,
-            makerToken: string,
-            makerTokenAmounts: BigNumber[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isString('takerToken', takerToken);
-            assert.isString('makerToken', makerToken);
-            assert.isArray('makerTokenAmounts', makerTokenAmounts);
+        takerToken: string,
+        makerToken: string,
+        makerTokenAmounts: BigNumber[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isString('takerToken', takerToken);
+        assert.isString('makerToken', makerToken);
+        assert.isArray('makerTokenAmounts', makerTokenAmounts);
         const functionSignature = 'sampleBuysFromEth2Dai(address,address,uint256[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [takerToken.toLowerCase(),
-            makerToken.toLowerCase(),
-            makerTokenAmounts
-            ]);
+                return self._strictEncodeArguments(functionSignature, [
+                    takerToken.toLowerCase(),
+                    makerToken.toLowerCase(),
+                    makerTokenAmounts,
+                ]);
             },
-        }
-    };
+        };
+    }
     /**
      * Sample buy quotes from an arbitrary on-chain liquidity provider.
-      * @param registryAddress Address of the liquidity provider registry contract.
-      * @param takerToken Address of the taker token (what to sell).
-      * @param makerToken Address of the maker token (what to buy).
-      * @param makerTokenAmounts Maker token buy amount for each sample.
-    * @returns takerTokenAmounts Taker amounts sold at each maker token         amount.
+     * @param registryAddress Address of the liquidity provider registry contract.
+     * @param takerToken Address of the taker token (what to sell).
+     * @param makerToken Address of the maker token (what to buy).
+     * @param makerTokenAmounts Maker token buy amount for each sample.
+     * @returns takerTokenAmounts Taker amounts sold at each maker token         amount.
      */
     public sampleBuysFromLiquidityProviderRegistry(
-            registryAddress: string,
-            takerToken: string,
-            makerToken: string,
-            makerTokenAmounts: BigNumber[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isString('registryAddress', registryAddress);
-            assert.isString('takerToken', takerToken);
-            assert.isString('makerToken', makerToken);
-            assert.isArray('makerTokenAmounts', makerTokenAmounts);
+        registryAddress: string,
+        takerToken: string,
+        makerToken: string,
+        makerTokenAmounts: BigNumber[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isString('registryAddress', registryAddress);
+        assert.isString('takerToken', takerToken);
+        assert.isString('makerToken', makerToken);
+        assert.isArray('makerTokenAmounts', makerTokenAmounts);
         const functionSignature = 'sampleBuysFromLiquidityProviderRegistry(address,address,address,uint256[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [registryAddress.toLowerCase(),
-            takerToken.toLowerCase(),
-            makerToken.toLowerCase(),
-            makerTokenAmounts
-            ]);
+                return self._strictEncodeArguments(functionSignature, [
+                    registryAddress.toLowerCase(),
+                    takerToken.toLowerCase(),
+                    makerToken.toLowerCase(),
+                    makerTokenAmounts,
+                ]);
             },
-        }
-    };
+        };
+    }
     /**
      * Sample buy quotes from Uniswap.
-      * @param takerToken Address of the taker token (what to sell).
-      * @param makerToken Address of the maker token (what to buy).
-      * @param makerTokenAmounts Maker token sell amount for each sample.
-    * @returns takerTokenAmounts Taker amounts sold at each maker token         amount.
+     * @param takerToken Address of the taker token (what to sell).
+     * @param makerToken Address of the maker token (what to buy).
+     * @param makerTokenAmounts Maker token sell amount for each sample.
+     * @returns takerTokenAmounts Taker amounts sold at each maker token         amount.
      */
     public sampleBuysFromUniswap(
-            takerToken: string,
-            makerToken: string,
-            makerTokenAmounts: BigNumber[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isString('takerToken', takerToken);
-            assert.isString('makerToken', makerToken);
-            assert.isArray('makerTokenAmounts', makerTokenAmounts);
+        takerToken: string,
+        makerToken: string,
+        makerTokenAmounts: BigNumber[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isString('takerToken', takerToken);
+        assert.isString('makerToken', makerToken);
+        assert.isArray('makerTokenAmounts', makerTokenAmounts);
         const functionSignature = 'sampleBuysFromUniswap(address,address,uint256[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [takerToken.toLowerCase(),
-            makerToken.toLowerCase(),
-            makerTokenAmounts
-            ]);
+                return self._strictEncodeArguments(functionSignature, [
+                    takerToken.toLowerCase(),
+                    makerToken.toLowerCase(),
+                    makerTokenAmounts,
+                ]);
             },
-        }
-    };
+        };
+    }
     /**
      * Sample sell quotes from Curve.
-      * @param curveAddress Address of the Curve contract.
-      * @param fromTokenIdx Index of the taker token (what to sell).
-      * @param toTokenIdx Index of the maker token (what to buy).
-      * @param takerTokenAmounts Taker token sell amount for each sample.
-    * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
+     * @param curveAddress Address of the Curve contract.
+     * @param fromTokenIdx Index of the taker token (what to sell).
+     * @param toTokenIdx Index of the maker token (what to buy).
+     * @param takerTokenAmounts Taker token sell amount for each sample.
+     * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
      */
     public sampleSellsFromCurve(
-            curveAddress: string,
-            fromTokenIdx: BigNumber,
-            toTokenIdx: BigNumber,
-            takerTokenAmounts: BigNumber[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isString('curveAddress', curveAddress);
-            assert.isBigNumber('fromTokenIdx', fromTokenIdx);
-            assert.isBigNumber('toTokenIdx', toTokenIdx);
-            assert.isArray('takerTokenAmounts', takerTokenAmounts);
+        curveAddress: string,
+        fromTokenIdx: BigNumber,
+        toTokenIdx: BigNumber,
+        takerTokenAmounts: BigNumber[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isString('curveAddress', curveAddress);
+        assert.isBigNumber('fromTokenIdx', fromTokenIdx);
+        assert.isBigNumber('toTokenIdx', toTokenIdx);
+        assert.isArray('takerTokenAmounts', takerTokenAmounts);
         const functionSignature = 'sampleSellsFromCurve(address,int128,int128,uint256[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [curveAddress.toLowerCase(),
-            fromTokenIdx,
-            toTokenIdx,
-            takerTokenAmounts
-            ]);
+                return self._strictEncodeArguments(functionSignature, [
+                    curveAddress.toLowerCase(),
+                    fromTokenIdx,
+                    toTokenIdx,
+                    takerTokenAmounts,
+                ]);
             },
-        }
-    };
+        };
+    }
     /**
      * Sample sell quotes from Eth2Dai/Oasis.
-      * @param takerToken Address of the taker token (what to sell).
-      * @param makerToken Address of the maker token (what to buy).
-      * @param takerTokenAmounts Taker token sell amount for each sample.
-    * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
+     * @param takerToken Address of the taker token (what to sell).
+     * @param makerToken Address of the maker token (what to buy).
+     * @param takerTokenAmounts Taker token sell amount for each sample.
+     * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
      */
     public sampleSellsFromEth2Dai(
-            takerToken: string,
-            makerToken: string,
-            takerTokenAmounts: BigNumber[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isString('takerToken', takerToken);
-            assert.isString('makerToken', makerToken);
-            assert.isArray('takerTokenAmounts', takerTokenAmounts);
+        takerToken: string,
+        makerToken: string,
+        takerTokenAmounts: BigNumber[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isString('takerToken', takerToken);
+        assert.isString('makerToken', makerToken);
+        assert.isArray('takerTokenAmounts', takerTokenAmounts);
         const functionSignature = 'sampleSellsFromEth2Dai(address,address,uint256[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [takerToken.toLowerCase(),
-            makerToken.toLowerCase(),
-            takerTokenAmounts
-            ]);
+                return self._strictEncodeArguments(functionSignature, [
+                    takerToken.toLowerCase(),
+                    makerToken.toLowerCase(),
+                    takerTokenAmounts,
+                ]);
             },
-        }
-    };
+        };
+    }
     /**
      * Sample sell quotes from Kyber.
-      * @param takerToken Address of the taker token (what to sell).
-      * @param makerToken Address of the maker token (what to buy).
-      * @param takerTokenAmounts Taker token sell amount for each sample.
-    * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
+     * @param takerToken Address of the taker token (what to sell).
+     * @param makerToken Address of the maker token (what to buy).
+     * @param takerTokenAmounts Taker token sell amount for each sample.
+     * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
      */
     public sampleSellsFromKyberNetwork(
-            takerToken: string,
-            makerToken: string,
-            takerTokenAmounts: BigNumber[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isString('takerToken', takerToken);
-            assert.isString('makerToken', makerToken);
-            assert.isArray('takerTokenAmounts', takerTokenAmounts);
+        takerToken: string,
+        makerToken: string,
+        takerTokenAmounts: BigNumber[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isString('takerToken', takerToken);
+        assert.isString('makerToken', makerToken);
+        assert.isArray('takerTokenAmounts', takerTokenAmounts);
         const functionSignature = 'sampleSellsFromKyberNetwork(address,address,uint256[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [takerToken.toLowerCase(),
-            makerToken.toLowerCase(),
-            takerTokenAmounts
-            ]);
+                return self._strictEncodeArguments(functionSignature, [
+                    takerToken.toLowerCase(),
+                    makerToken.toLowerCase(),
+                    takerTokenAmounts,
+                ]);
             },
-        }
-    };
+        };
+    }
     /**
      * Sample sell quotes from an arbitrary on-chain liquidity provider.
-      * @param registryAddress Address of the liquidity provider registry contract.
-      * @param takerToken Address of the taker token (what to sell).
-      * @param makerToken Address of the maker token (what to buy).
-      * @param takerTokenAmounts Taker token sell amount for each sample.
-    * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
+     * @param registryAddress Address of the liquidity provider registry contract.
+     * @param takerToken Address of the taker token (what to sell).
+     * @param makerToken Address of the maker token (what to buy).
+     * @param takerTokenAmounts Taker token sell amount for each sample.
+     * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
      */
     public sampleSellsFromLiquidityProviderRegistry(
-            registryAddress: string,
-            takerToken: string,
-            makerToken: string,
-            takerTokenAmounts: BigNumber[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isString('registryAddress', registryAddress);
-            assert.isString('takerToken', takerToken);
-            assert.isString('makerToken', makerToken);
-            assert.isArray('takerTokenAmounts', takerTokenAmounts);
+        registryAddress: string,
+        takerToken: string,
+        makerToken: string,
+        takerTokenAmounts: BigNumber[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isString('registryAddress', registryAddress);
+        assert.isString('takerToken', takerToken);
+        assert.isString('makerToken', makerToken);
+        assert.isArray('takerTokenAmounts', takerTokenAmounts);
         const functionSignature = 'sampleSellsFromLiquidityProviderRegistry(address,address,address,uint256[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [registryAddress.toLowerCase(),
-            takerToken.toLowerCase(),
-            makerToken.toLowerCase(),
-            takerTokenAmounts
-            ]);
+                return self._strictEncodeArguments(functionSignature, [
+                    registryAddress.toLowerCase(),
+                    takerToken.toLowerCase(),
+                    makerToken.toLowerCase(),
+                    takerTokenAmounts,
+                ]);
             },
-        }
-    };
+        };
+    }
     /**
      * Sample sell quotes from Uniswap.
-      * @param takerToken Address of the taker token (what to sell).
-      * @param makerToken Address of the maker token (what to buy).
-      * @param takerTokenAmounts Taker token sell amount for each sample.
-    * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
+     * @param takerToken Address of the taker token (what to sell).
+     * @param makerToken Address of the maker token (what to buy).
+     * @param takerTokenAmounts Taker token sell amount for each sample.
+     * @returns makerTokenAmounts Maker amounts bought at each taker token         amount.
      */
     public sampleSellsFromUniswap(
-            takerToken: string,
-            makerToken: string,
-            takerTokenAmounts: BigNumber[],
-    ): ContractFunctionObj<BigNumber[]
-> {
-        const self = this as any as ERC20BridgeSamplerContract;
-            assert.isString('takerToken', takerToken);
-            assert.isString('makerToken', makerToken);
-            assert.isArray('takerTokenAmounts', takerTokenAmounts);
+        takerToken: string,
+        makerToken: string,
+        takerTokenAmounts: BigNumber[],
+    ): ContractFunctionObj<BigNumber[]> {
+        const self = (this as any) as ERC20BridgeSamplerContract;
+        assert.isString('takerToken', takerToken);
+        assert.isString('makerToken', makerToken);
+        assert.isArray('takerTokenAmounts', takerTokenAmounts);
         const functionSignature = 'sampleSellsFromUniswap(address,address,uint256[])';
 
         return {
-            async callAsync(
-                callData: Partial<CallData> = {},
-                defaultBlock?: BlockParam,
-            ): Promise<BigNumber[]
-            > {
+            async callAsync(callData: Partial<CallData> = {}, defaultBlock?: BlockParam): Promise<BigNumber[]> {
                 BaseContract._assertCallParams(callData, defaultBlock);
-                const rawCallResult = await self._performCallAsync({ ...callData, data: this.getABIEncodedTransactionData() }, defaultBlock);
+                const rawCallResult = await self._performCallAsync(
+                    { ...callData, data: this.getABIEncodedTransactionData() },
+                    defaultBlock,
+                );
                 const abiEncoder = self._lookupAbiEncoder(functionSignature);
                 BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
-                return abiEncoder.strictDecodeReturnValue<BigNumber[]
-            >(rawCallResult);
+                return abiEncoder.strictDecodeReturnValue<BigNumber[]>(rawCallResult);
             },
             getABIEncodedTransactionData(): string {
-                return self._strictEncodeArguments(functionSignature, [takerToken.toLowerCase(),
-            makerToken.toLowerCase(),
-            takerTokenAmounts
-            ]);
+                return self._strictEncodeArguments(functionSignature, [
+                    takerToken.toLowerCase(),
+                    makerToken.toLowerCase(),
+                    takerTokenAmounts,
+                ]);
             },
-        }
-    };
-
-
+        };
+    }
 
     constructor(
         address: string,
@@ -1160,9 +1161,17 @@ public static async deployFrom0xArtifactAsync(
         logDecodeDependencies?: { [contractName: string]: ContractAbi },
         deployedBytecode: string | undefined = ERC20BridgeSamplerContract.deployedBytecode,
     ) {
-        super('ERC20BridgeSampler', ERC20BridgeSamplerContract.ABI(), address, supportedProvider, txDefaults, logDecodeDependencies, deployedBytecode);
+        super(
+            'ERC20BridgeSampler',
+            ERC20BridgeSamplerContract.ABI(),
+            address,
+            supportedProvider,
+            txDefaults,
+            logDecodeDependencies,
+            deployedBytecode,
+        );
         classUtils.bindAll(this, ['_abiEncoderByFunctionSignature', 'address', '_web3Wrapper']);
-ERC20BridgeSamplerContract.ABI().forEach((item, index) => {
+        ERC20BridgeSamplerContract.ABI().forEach((item, index) => {
             if (item.type === 'function') {
                 const methodAbi = item as MethodAbi;
                 this._methodABIIndex[methodAbi.name] = index;
